moves the query result to class
